Avoid 'conversion may change value' cases in advancedfilter.cpp
Avoid two warnings in advancedfilter.cpp, confirming conversions to `int`
because a non-Krusader function that is used (`setValue(int value)`) only
accepts that data type.

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/57

+++ krusader/Filter/advancedfilter.cpp
@@ -550,11 +550,11 @@ bool AdvancedFilter::getSettings(FilterSettings &s)
 void AdvancedFilter::applySettings(const FilterSettings &s)
 {
     minSizeEnabled->setChecked(s.minSizeEnabled);
-    minSizeAmount->setValue(s.minSize.amount);
+    minSizeAmount->setValue(static_cast<int>(s.minSize.amount));
     minSizeType->setCurrentIndex(s.minSize.unit);
 
     maxSizeEnabled->setChecked(s.maxSizeEnabled);
-    maxSizeAmount->setValue(s.maxSize.amount);
+    maxSizeAmount->setValue(static_cast<int>(s.maxSize.amount));
     maxSizeType->setCurrentIndex(s.maxSize.unit);
 
     if (s.modifiedBetweenEnabled)
